+++ ZooManagementSystem/src/main/java/com/example/zoo/repository/DongVatRepository.java
@@ -13,7 +13,8 @@ public interface DongVatRepository extends JpaRepository<DongVat, Long> {
     boolean existsByMaChuong(String maChuong);
     List<DongVat> findByLoai(String loai);
     List<DongVat> findBySucKhoe(String sucKhoe);
-    List<DongVat> findByTuoiBetween(int minTuoi, int maxTuoi);
+    
+    List<DongVat> findByTuoiBetween(double minTuoi, double maxTuoi);
     
     DongVat findByTen(String ten);
     List<DongVat> findByTenContainingIgnoreCase(String ten);
